requirements.txt
PyGithub
python-dotenv
pandas
pyarrow
boto3
pqdm
optuna
xgboost
scikit-learn
moto[s3]==4.2.13
pytest
